[PATCH][BLKPARSE/BLKTRACE] too small arrays for file names
In our experiments blktrace/blkparse file names encode a lot of infomation about
the particular experiment.  We noticed that for long enough file names blkparse
does not work.

The reason is that per_cpu_info->fname[] is of 128 bytes. As a result, in
setup_file() function only part of the file name gets to ->fname[]. Then stat()
fails and we exit the function. Notice, that no error is printed in this case.

In the following patch ->fname[] size is increased to POSXI defined PATH_MAX.

Signed-off-by: Vasily Tarasov <tarasov@vasily.name>

(Submitted to upstream,  but not integrated yet)