Remove environment variable STRACE2DS
When a user installs this library in a directory other than
/usr/local/strace2ds, we still need to find the xml and table files.
Previously, we relied on the user setting an environment variable that
contained the path where strace2ds was installed. Now, we check for the
files with the relative path ../lib/strace2ds. This needs to be made
more robust, while still avoiding a kludgy environment variable.