Create a correlation analysis
CorrelationAnalysisModule.hpp has been added to the include
directory. It containes a class CorrelationAnalysisModule.
The corresponding .cpp file has been added to the src/analysis
directory. SystemCallTraceReplayer.hpp and .cpp were changed
to run the correlation analysis. src/analysis/CMakeLists.txt
was updated accordingly.

Currently the CorrelationAnalysisModule returns the names
of all systemcalls in the .ds file and the number of times
each of the system calls appears.